# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task manager",
    "description": "Created by JustDenis\n",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/justdenis7/task-manager-api/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "http://localhost:3000/",
      "description": "Base url"
    }
  ],
...
  "paths": {
...
    "/sprints/create": {
      "post": {
        "tags": [
          "Sprints"
        ],
        "summary": "Create sprint",
        "operationId": "createSprint",
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateSprint"
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/sprints/{projectId}/{sprintId}/remove": {
      "delete": {
        "tags": [
          "Sprints"
        ],
        "summary": "Remove sprint",
        "operationId": "removeSprint",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "Project id",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sprintId",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/sprints/{sprintId}": {
      "get": {
        "tags": [
          "Sprints"
        ],
        "summary": "Get current sprint",
        "operationId": "currentSprint",
        "parameters": [
          {
            "name": "sprintId",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentSprint"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/tasks/create": {
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create task",
        "operationId": "createTask",
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateTask"
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/tasks/{sprintId}/{taskId}/remove": {
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Remove task",
        "operationId": "removeTask",
        "parameters": [
          {
            "name": "sprintId",
            "in": "path",
            "description": "Sprint id",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "in": "path",
            "description": "Task id",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSprint": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "startAt": {
            "type": "string"
          },
          "finishedAt": {
            "type": "string"
          }
        }
      },
      "CurrentSprint": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CurrentSprintItem"
            }
          },
          "isOwner": {
            "type": "boolean"
          }
        }
      },
      "CurrentSprintItem": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "CreateTask": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "plannedTime": {
            "type": "number"
          }
        }
      },
...
    },
    "requestBodies": {
...
      "CreateSprint": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateSprint"
            }
          }
        }
      },
      "CreateTask": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/CreateTask"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}